Remove Configs reducer
