+++ src/js/api/configurations/configurations-api.ts
@@ -1,16 +1,42 @@
 import ConfigPropValues from "src/js/state/ConfigPropValues"
-import Configs, {Config} from "src/js/state/Configs"
 import {AppDispatch, State} from "src/js/state/types"
 
+export type ConfigItemType = "file" | "string" | "directory" | "boolean" // | "number" | "boolean"
+
+export type ConfigItem = {
+  name: string
+  type: ConfigItemType
+  label: string
+  helpLink?: {
+    label: string
+    url: string
+  }
+  command?: string
+  defaultValue?: string | boolean
+  enum?: string[]
+}
+
+export type Config = {
+  name: string
+  title: string
+  properties: {[configItemName: string]: ConfigItem}
+}
+
 export class ConfigurationsApi {
+  configs: Config[] = []
+
   constructor(private dispatch: AppDispatch, private getState: () => State) {}
 
+  get all() {
+    return this.configs
+  }
+  k
   get(configName, propName): any {
     return ConfigPropValues.get(configName, propName)(this.getState())
   }
 
   add(config: Config) {
-    this.dispatch(Configs.set(config))
+    this.configs.push(config)
     // Set default values if there are any unset values
     for (let name in config.properties) {
       const prop = config.properties[name]
