+++ src/js/components/Preferences/usePreferencesForm.ts
@@ -1,7 +1,6 @@
-import {useSelector} from "react-redux"
+import {useBrimApi} from "src/app/core/context"
 import {useDispatch} from "src/app/core/state"
 import ConfigPropValues from "src/js/state/ConfigPropValues"
-import Configs from "src/js/state/Configs"
 import {FormConfig, FormFieldConfig} from "../../brim/form"
 import {executeCommand} from "../../flows/executeCommand"
 import lib from "../../lib"
@@ -16,7 +15,8 @@ const checkFile = (path) => {
 
 export const useConfigsForm = (): FormConfig => {
   const dispatch = useDispatch()
-  const configs = useSelector(Configs.all)
+  const api = useBrimApi()
+  const configs = api.configs.all
   const formConfig: FormConfig = {}
   configs.forEach((config) => {
     Object.values(config.properties).forEach((prop) => {
